Add allowing book details entering
* Change some layout
* Change book activity to book details activity
* rename corresponding files to match details naming
* Add reusable toolbar layout

+++ app/src/main/java/org/thelinuxmotion/apps/booktracker/bookinfo/BookReadingDetails.java
@@ -17,7 +17,7 @@ public class BookReadingDetails {
     public @NonNull String mIsbn;
 
     @ColumnInfo(name = "day")
-    public long mDay;
+    public int mDay;
 
     @ColumnInfo(name = "date")
     public long mDateTime;
